### 2022-01-21 #4086 @kyleniemeyer
Hi @bkmi, is your JOSS paper located in a particular branch? Our bot (and I) can't seem to find it.
