### 2022-01-24 #4086 @fabian-s
Sorry for the late reply -- I'd rather not, since I don't know Python and the problem this package solves is outside my area of expertise.
