+++ paper/paper.bib
@@ -388,7 +388,8 @@
   number={3},
   pages={4440--4458},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/stz1960}
 }
 
 @article{alsing2018massive,
@@ -399,7 +400,8 @@
   number={3},
   pages={2874--2885},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/sty819}
 }
 
 @article{alsing2019nuisance,
@@ -410,7 +412,8 @@
   number={4},
   pages={5093--5103},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/stz1900}
 }
 
 %%%%%%%%%%%%%%%%%%%%%%% Simulation-based inference algorithms - likelihood ratio %%%%%%%%%%%%%%%%%%%%%%%
@@ -430,7 +433,8 @@
   number={1},
   pages={218--227},
   year={2014},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/sta4.56}
 }
 
 @article{thomas2016likelihood,
@@ -438,7 +442,8 @@
   author={Thomas, Owen and Dutta, Ritabrata and Corander, Jukka and Kaski, Samuel and Gutmann, Michael U and others},
   journal={Bayesian Analysis},
   year={2016},
-  publisher={International Society for Bayesian Analysis}
+  publisher={International Society for Bayesian Analysis},
+  doi={10.1214/20-ba1238}
 }
 
 @article{gutmann2018likelihood,
@@ -449,7 +454,8 @@
   number={2},
   pages={411--425},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11222-017-9738-6}
 }
 
 @article{Cranmer2015,
@@ -473,7 +479,8 @@
 	year      =  2006,
 	keywords  = "Bayesian computation; evidence; marginal likelihood; algorithm;
 	nest; annealing; phase change; model selection",
-	language  = "en"
+	language  = "en",
+	doi 	  = "10.1214/06-ba127",
 }
 
 
@@ -493,7 +500,8 @@
 	pages     = "449--463",
 	month     =  jan,
 	year      =  2008,
-	language  = "en"
+	language  = "en",
+	doi 	  = "10.1111/j.1365-2966.2007.12353.x"
 }
 
 
@@ -525,7 +533,8 @@
 	journal  = "Mon. Not. Roy. Astron. Soc. 398: 1601-1614,2009",
 	month    =  sep,
 	year     =  2008,
-	keywords = "astro-ph"
+	keywords = "astro-ph",
+	doi		 = "10.1111/j.1365-2966.2009.14548.x"
 }
 
 
@@ -542,7 +551,8 @@
 	pages     = "4384--4398",
 	month     =  sep,
 	year      =  2015,
-	language  = "en"
+	language  = "en",
+	doi		  = "10.1093/mnras/stv1911"
 }
 
 
@@ -568,7 +578,8 @@
   number={5},
   pages={891--913},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11222-018-9844-0}
 }
 
 %%%%%%%%%%%%% ABC %%%%%%%%%%%%%
@@ -586,7 +597,8 @@
   volume={12},
   pages={66--104},
   year={2018},
[…]
