### 2022-01-25 #4086 @bkmi
May I suggest @dfm as editor? If he's willing, of course. Although generally applicable to simulation-based inference, our software is utilized in astrophysics papers.
