+++ swyft/lightning/components.py
@@ -20,6 +20,7 @@ import swyft
 from dataclasses import dataclass
 import pytorch_lightning as pl
 from torch.nn import functional as F
+from swyft.inference.marginalratioestimator import get_ntrain_nvalid
 
 from typing import (
     Callable,
@@ -146,14 +147,15 @@ class RectangleBound:
 def get_1d_rect_bounds(samples, th = 1e-6):
     bounds = {}
     for k, v in samples.items():
-        w = v.ratios
+        r = v.ratios
+        r = r - r.max(axis=0).values  # subtract peak
         p = v.values
         #w = v[..., 0]
         #p = v[..., 1]
         all_max = p.max(dim=0).values
         all_min = p.min(dim=0).values
-        constr_min = torch.where(w > np.log(th), p, all_max).min(dim=0).values
-        constr_max = torch.where(w > np.log(th), p, all_min).max(dim=0).values
+        constr_min = torch.where(r > np.log(th), p, all_max).min(dim=0).values
+        constr_max = torch.where(r > np.log(th), p, all_min).max(dim=0).values
         #bound = torch.stack([constr_min, constr_max], dim = -1)
         bound = RectangleBound(constr_min, constr_max)
         bounds[k] = bound
@@ -166,6 +168,13 @@ def append_randomized(z):
     z = torch.cat([z, z[n+idx], z[idx]])
     return z
 
+def append_nonrandomized(z):
+    assert len(z)%2 == 0, "Cannot expand odd batch dimensions."
+    n = len(z)//2
+    idx = np.arange(n)
+    z = torch.cat([z, z[n+idx], z[idx]])
+    return z
+
 def valmap(m, d):
     return {k: m(v) for k, v in d.items()}
 
@@ -175,10 +184,13 @@ class SwyftModule(pl.LightningModule):
         self._predict_condition_x = {}
         self._predict_condition_z = {}
         
+    def on_train_start(self):
+        self.logger.log_hyperparams(self.hparams, {"hp/KL-div": 0, "hp/JS-div": 0})
+        
     def configure_optimizers(self):
         optimizer = torch.optim.Adam(self.parameters(), lr=1e-3)
         #scheduler = torch.optim.lr_scheduler.StepLR(optimizer, 30)
-        lr_scheduler = {"scheduler": torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer), "monitor": "val_loss"}
+        lr_scheduler = {"scheduler": torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience = 2, factor = 0.25), "monitor": "val_loss"}
         return dict(optimizer = optimizer, lr_scheduler = lr_scheduler)
 
     def _log_ratios(self, x, z):
@@ -188,23 +200,40 @@ class SwyftModule(pl.LightningModule):
     
     def validation_step(self, batch, batch_idx):
         loss = self._calc_loss(batch, batch_idx)
-        self.log("val_loss", loss)
+        self.log("val_loss", loss, prog_bar=True)
         return loss
-        
-    def _calc_loss(self, batch, batch_idx):
+
+    def _calc_loss(self, batch, batch_idx, randomized = True):
         x, z = batch
-        z = valmap(append_randomized, z)
+        if randomized:
+            z = valmap(append_randomized, z)
+        else:
+            z = valmap(append_nonrandomized, z)
         log_ratios = self._log_ratios(x, z)
         nbatch = len(log_ratios)//2
         y = torch.zeros_like(log_ratios)
         y[:nbatch, ...] = 1
         loss = F.binary_cross_entropy_with_logits(log_ratios, y, reduce = False)
-        loss = loss.sum()
+        loss = loss.sum()/nbatch
[…]
