### 2022-02-23 fe9f5cf
Switch on online normalization
