+++ swyft/lightning/components_v1.py
@@ -1,0 +1,485 @@
+import pylab as plt
+import numpy as np
+import skimage.transform
+import skimage.restoration
+from tqdm import tqdm
+import skimage.filters
+import torch
+import torch.nn as nn
+import typing as tp
+from pyrofit.utils import pad_dims
+from torch.nn.functional import grid_sample
+from pyro import distributions as dist
+from pyrofit.lensing.utils import get_meshgrid
+from pyrofit.lensing.lenses import SPLELens
+from pyrofit.lensing.sources import SersicSource, AnalyticSource
+from pyrofit.utils.torchutils import _mid_many, unravel_index
+from pyrofit.utils import kNN
+from fft_conv_pytorch import fft_conv, FFTConv2d
+import swyft
+from dataclasses import dataclass
+import pytorch_lightning as pl
+from torch.nn import functional as F
+from swyft.inference.marginalratioestimator import get_ntrain_nvalid
+
+from typing import (
+    Callable,
+    Dict,
+    Hashable,
+    Optional,
+    Sequence,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
+
+import numpy as np
+import torch
+import torch.nn as nn
+from toolz.dicttoolz import valmap
+from torch.utils.data import DataLoader, Dataset, random_split
+
+import swyft
+import swyft.utils
+from swyft.saveable import StateDictSaveable, StateDictSaveableType
+from swyft.types import Array, Device, MarginalIndex, MarginalToArray, ObsType, PathType
+from swyft.utils.array import array_to_tensor, dict_array_to_tensor
+from swyft.utils.marginals import tupleize_marginal_indices
+
+
+
+class RatioEstimatorGaussian1d(torch.nn.Module):
+    def __init__(self, momentum = 0.1):
+        super().__init__()
+        self.momentum = momentum        
+        self.x_mean = None
+        self.z_mean = None
+        self.x_var = None
+        self.z_var = None
+        self.xz_cov = None
+        
+    def forward(self, x: torch.Tensor, z: torch.Tensor) -> torch.Tensor:
+        """2-dim Gaussian approximation to marginals and joint, assuming (B, N)."""
+        if self.training or self.x_mean is None:
+            # Covariance estimates must be based on joined samples only
+            # NOTE: This makes assumptions about the structure of mini batches during training (J, M, M, J, J, M, M, J, ...)
+            # TODO: Change to (J, M, J, M, J, M, ...) in the future
+            batch_size = len(x)
+            #idx = np.array([[i, i+3] for i in np.arange(0, batch_size, 4)]).flatten() 
+            idx = np.arange(batch_size//2)  # TODO: Assuming (J, J, J, J, M, M, M, M) etc
+            
+            # Estimation w/o Bessel's correction, using simple MLE estimate (https://en.wikipedia.org/wiki/Estimation_of_covariance_matrices)
+            x_mean_batch = x[idx].mean(dim=0).detach()
+            z_mean_batch = z[idx].mean(dim=0).detach()
+            x_var_batch = ((x[idx]-x_mean_batch)**2).mean(dim=0).detach()
+            z_var_batch = ((z[idx]-z_mean_batch)**2).mean(dim=0).detach()
+            xz_cov_batch = ((x[idx]-x_mean_batch)*(z[idx]-z_mean_batch)).mean(dim=0).detach()
+            
+            # Momentum-based update rule
+            momentum = self.momentum
+            self.x_mean = x_mean_batch if self.x_mean is None else (1-momentum)*self.x_mean + momentum*x_mean_batch
+            self.x_var = x_var_batch if self.x_var is None else (1-momentum)*self.x_var + momentum*x_var_batch
[…]
