### 2022-02-24 790b746
Clean up lightning components
