+++ swyft/lightning/components.py
@@ -1,27 +1,5 @@
-import pylab as plt
-import numpy as np
-import skimage.transform
-import skimage.restoration
-from tqdm import tqdm
-import skimage.filters
-import torch
-import torch.nn as nn
-import typing as tp
-from pyrofit.utils import pad_dims
-from torch.nn.functional import grid_sample
-from pyro import distributions as dist
-from pyrofit.lensing.utils import get_meshgrid
-from pyrofit.lensing.lenses import SPLELens
-from pyrofit.lensing.sources import SersicSource, AnalyticSource
-from pyrofit.utils.torchutils import _mid_many, unravel_index
-from pyrofit.utils import kNN
-from fft_conv_pytorch import fft_conv, FFTConv2d
-import swyft
 from dataclasses import dataclass, field
-import pytorch_lightning as pl
-from torch.nn import functional as F
-from swyft.inference.marginalratioestimator import get_ntrain_nvalid
-
+from toolz.dicttoolz import valmap
 from typing import (
     Callable,
     Dict,
@@ -33,152 +11,81 @@ from typing import (
     TypeVar,
     Union,
 )
-
 import numpy as np
 import torch
 import torch.nn as nn
-from toolz.dicttoolz import valmap
-from torch.utils.data import DataLoader, Dataset, random_split
-
+from torch.nn import functional as F
+from torch.utils.data import random_split
+import pytorch_lightning as pl
+from tqdm import tqdm
 import swyft
 import swyft.utils
-from swyft.saveable import StateDictSaveable, StateDictSaveableType
-from swyft.types import Array, Device, MarginalIndex, MarginalToArray, ObsType, PathType
-from swyft.utils.array import array_to_tensor, dict_array_to_tensor
-from swyft.utils.marginals import tupleize_marginal_indices
+from swyft.inference.marginalratioestimator import get_ntrain_nvalid
 
 
+#################################
+# Swyft lightning main components
+#################################
 
-class RatioEstimatorGaussian1d(torch.nn.Module):
-    def __init__(self, momentum = 0.1):
+class SwyftModel:
+    def _simulate(self, N, bounds = None, effective_prior = None):
+        # TODO: Include conditional priors
+        prior_samples = self.prior(N, bounds = bounds)
+        if effective_prior:
+            for k in effective_prior.keys():
+                samples = effective_prior[k].sample(N)
+                prior_samples[k] = samples[k]
+        simulated_samples = dictstoremap(self.slow, prior_samples)
+        samples_tot = dict(**prior_samples, **simulated_samples)
+        return SampleStore(samples_tot)
+    
+    def sample(self, N, bounds = None, effective_prior = None):
+        sims = self._simulate(N, bounds = bounds, effective_prior = effective_prior)
+        data = dictstoremap(self.fast, sims)
+        out = dict(**data, **sims)
+        return SampleStore(out)
+    
+    def noise(self, S):
+        S = S.copy()
+        D = self.fast(S)
+        S.update(D)
+        return S#dict(**D, **S)
+    
+    def __call__(self, S):
+        D = self.slow(S)
+        D = dict(**D, **S)
+        E = self.fast(D)
+        return dict(**D, **E)
+
+
+class SwyftDataModule(pl.LightningDataModule):
+    def __init__(self, store, model = None, batch_size: int = 32, validation_percentage = 0.2, manual_seed = None, train_multiply = 10 ):
         super().__init__()
-        self.momentum = momentum        
-        self.x_mean = None
-        self.z_mean = None
-        self.x_var = None
-        self.z_var = None
-        self.xz_cov = None
-        
[…]
