### 2022-02-25 #4205 @editorialbot
[REVIEW]: swyft: Truncated Marginal Neural Ratio Estimation in Python
**Submitting author:** @bkmi (<a href="http://orcid.org/0000-0003-0387-8727">Benjamin Miller</a>)
**Repository:** https://github.com/undark-lab/swyft
**Branch with paper.md** (empty if default branch): joss-submission-do-not-delete
**Version:** v0.3.2
