### 2022-02-25 #4205 @pdebuyl
Hi @mattpitkin @olgadoronina please read full the instructions at https://github.com/openjournals/joss-reviews/issues/4205#issue-1150402557 , they contain all the necessary information. I remain available in any case.
