### 2022-02-25 #4205 @mattpitkin
@bkmi I've just installed version 0.2.1 of swyft from PyPI. I'm trying to run the [Quickstart guide](https://swyft.readthedocs.io/en/latest/notebooks/Quickstart.html), but on line:

```python
marginal_indices_1d, marginal_indices_2d = swyft.utils.get_corner_marginal_indices(n_parameters)
```

it gives the error:

```python
AttributeError: module 'swyft.utils' has no attribute 'get_corner_marginal_indices'
```

I assume this is because v0.2.1 isn't up-to-date. Could you push the latest release to PyPI?
