### 2022-02-25 #4205 @mattpitkin
The paper is very nice. I only have a minor suggestion, which is due to my gravitational-wave bias. Seeing as rapid inference for gravitational waves is mentioned, it might be worth citing some of the related work in, e.g., [Gabbard _et al_](https://arxiv.org/abs/1909.06296), [Dax _et al_](https://ui.adsabs.harvard.edu/abs/2021PhRvL.127x1103D/abstract), [Chua & Vallisneri](https://ui.adsabs.harvard.edu/abs/2020PhRvL.124d1102C/abstract)  and (a paper including one the co-authors of swyft) [Delaunay _et al_](https://ui.adsabs.harvard.edu/abs/2020arXiv201012931D/abstract).
