### 2022-02-28 7d17f25
Update setup.cfg
