### 2022-02-28 8bb9176
Update setup.cfg
