+++ setup.cfg
@@ -3,7 +3,7 @@ name = swyft
 # version = We are using setuptools_scm, see https://github.com/pypa/setuptools_scm
 description = Truncated Marginal Neural Ratio Estimation with an inhomogeneous poisson point process cache.
 long_description = file: README.rst
-long_description_content_type = text/markdown
+long_description_content_type = text/x-rst
 url = https://github.com/undark-lab/swyft
 license = MIT License
 classifiers =
