### 2022-02-28 #4205 @bkmi
the `pip install swyft` should now be addressed and the user would be able to access the documentation and examples in notebooks.

I still need to make the changes w.r.t. the citations!
