+++ paper/paper.bib
@@ -722,3 +722,47 @@
 	year      =  1970,
 	doi		  =  "10.1093/biomet/57.1.97"
 }
+
+%%%%%%%%%%%%% applications %%%%%%%%%%%%%
+@article{gabbard2022bayesian,
+  title={Bayesian parameter estimation using conditional variational autoencoders for gravitational-wave astronomy},
+  author={Gabbard, Hunter and Messenger, Chris and Heng, Ik Siong and Tonolini, Francesco and Murray-Smith, Roderick},
+  journal={Nature Physics},
+  volume={18},
+  number={1},
+  pages={112--117},
+  year={2022},
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41567-021-01425-7}
+}
+
+@article{dax2021real,
+  title={Real-Time Gravitational Wave Science with Neural Posterior Estimation},
+  author={Dax, Maximilian and Green, Stephen R and Gair, Jonathan and Macke, Jakob H and Buonanno, Alessandra and Sch{\"o}lkopf, Bernhard},
+  journal={Physical review letters},
+  volume={127},
+  number={24},
+  pages={241103},
+  year={2021},
+  publisher={APS},
+  doi={10.1103/PhysRevLett.127.241103}
+}
+
+@article{chua2020learning,
+  title={Learning Bayesian posteriors with neural networks for gravitational-wave inference},
+  author={Chua, Alvin JK and Vallisneri, Michele},
+  journal={Physical review letters},
+  volume={124},
+  number={4},
+  pages={041102},
+  year={2020},
+  publisher={APS},
+  doi={10.1103/PhysRevLett.124.041102},
+}
+
+@article{delaunoy2020lightning,
+  title={Lightning-fast gravitational wave parameter inference through neural amortization},
+  author={Delaunoy, Arnaud and Wehenkel, Antoine and Hinderer, Tanja and Nissanke, Samaya and Weniger, Christoph and Williamson, Andrew R and Louppe, Gilles},
+  journal={Machine Learning and the Physical Sciences: Workshop at the 34th Conference on Neural Information Processing Systems (NeurIPS)},
+  year={2020}
+}
