### 2022-03-01 #4205 @bkmi
I added some citations!

p.s. due to some release issues the version that should be reviewed is actually `v0.3.1.post0`.
