### 2022-03-02 #4205 @mattpitkin
@bkmi I see that the version generation is now via setuptools_scm, which is great. I can get the version number by doing:

```python
from swyft.__version__ import version
print(version)
```

However, it would be great if the version number could be obtained just by using:

```python
import swyft
print(swyft.__version__)
```

You can achieve this by adding the following to the `swyft/__init__.py` file:

```python
try:
    from .__version__ import version as __version__
except ModuleNotFoundError:
    __version__ = ""
```
