### 2022-03-02 4d66257
added swyft.__version__ symantics. fixed import errors on some __all__ items.
