### 2022-03-02 #4205 @bkmi
Nice tip! I didn't know about that and I will use it in all my future projects!

I have made the change (along with fixing a `from swyft import *` bug which I found) in [this commit](https://github.com/undark-lab/swyft/commit/4d662577f8b75bcd7f8e614e0fad4def9815d15e). Would you like me to make an new release for it or should I wait until I incorporate other changes from this review process?
