### 2022-03-03 #4205 @mattpitkin
I have successfully installed and run several of the examples and confirmed that I achieve consistent results. I'm happy to sign-off swyft as reviewed.

I have a couple of questions unrelated to the review, one's a stylistic one and the other's more scientific:

* would it be possible in the violin plots to pass in the parameter names as labels? From looking at the `hist1d` code this looks possible, but not necessarily for the violin plots.
* (I should probably read the related papers, but...) given that swyft approximates the likelihood to evidence ratio, it is possible in any way to extract the evidence value too?
