### 2022-03-04 648460c
Add simple ZarrStore
