### 2022-03-04 cecc034
Add process lock to zarr store
