### 2022-03-04 7e65416
Make lr adjustable
