+++ swyft/lightning/components.py
@@ -92,16 +92,17 @@ class SwyftDataModule(pl.LightningDataModule):
 
 
 class SwyftModule(pl.LightningModule):
-    def __init__(self):
+    def __init__(self, lr = 1e-2):
         super().__init__()
         self._predict_condition_x = {}
         self._predict_condition_z = {}
+        self._lr = lr
         
     def on_train_start(self):
         self.logger.log_hyperparams(self.hparams, {"hp/KL-div": 0, "hp/JS-div": 0})
         
     def configure_optimizers(self):
-        optimizer = torch.optim.Adam(self.parameters(), lr=1e-4)
+        optimizer = torch.optim.Adam(self.parameters(), lr=self._lr)
         #scheduler = torch.optim.lr_scheduler.StepLR(optimizer, 30)
         lr_scheduler = {"scheduler": torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience = 2, factor = 0.25), "monitor": "val_loss"}
         return dict(optimizer = optimizer, lr_scheduler = lr_scheduler)
