### 2022-03-08 #4205 @bkmi
> I have successfully installed and run several of the examples and confirmed that I achieve consistent results. I'm happy to sign-off swyft as reviewed.
>
> I have a couple of questions unrelated to the review, one's a stylistic one and the other's more scientific:
>
> * would it be possible in the violin plots to pass in the parameter names as labels? From looking at the `hist1d` code this looks possible, but not necessarily for the violin plots.

This is a good point! I will make an issue and include this in the next release.

> * (I should probably read the related papers, but...) given that swyft approximates the likelihood to evidence ratio, it is possible in any way to extract the evidence value too?

It isn't really obvious how to do this to me, except through some kind of hacks... If you have the likelihood for a problem (swyft aims to assist in problems in which the likelihood is unknown), in principle you can estimate the likelihood-to-evidence ratio and then compute `likelihood * (1 / likelihood-to-evidence ratio)`. This is actually done as a way to show that the method works in https://arxiv.org/abs/1903.04057, but I wouldn't necessarily trust it for other problems. If you have the likelihood, you can also use nested sampling methods which are perhaps more trustworthy.
