### 2022-03-08 undark-lab/swyft#104 @bkmi
improve violin plot parameter labels
As mentioned in our JOSS review here https://github.com/openjournals/joss-reviews/issues/4205#issuecomment-1057974381 ...

It would be nice to pass parameter names to violin plots!
