### 2022-03-10 2c0ed51
added violin labels like #104, fixed a typing issue in hist1d
