### 2022-03-10 f984045
Handling of lr hyperparameter
