### 2022-03-10 7715713
Simpler arguments to network
