### 2022-03-10 66f909d
Log complete config file as hparams
