### 2022-03-10 38f15af
Add basic lightning script examples
