### 2022-03-14 1e7c2a0
Add get_shapes to simulator
