### 2022-03-15 0ce2673
Removed hydra again
