### 2022-03-16 06c4c05
Minor changes to simulator object
