### 2022-03-16 7177b0f
Add convenience fncts
