### 2022-03-16 f9d0d4b
Minor convenience improvements
