### 2022-03-16 2f59a47
Add get_sample_store function
