+++ swyft/lightning/components.py
@@ -65,12 +65,13 @@ class SwyftTrace(dict):
         if self.covers_targets:
             raise CoversTargetException
 
-class SwyftModelForward:
-    offline = True
-    online = True
 
+class SwyftModelForward:
     def forward(self, trace):
         raise NotImplementedError
+
+    def forward_afterburner(self, trace):
+        pass
 
     def _run(self, targets = None, conditions = {}, overwrite = False):
         trace = SwyftTrace(targets, conditions, overwrite = overwrite)
@@ -78,23 +79,15 @@ class SwyftModelForward:
             return trace
         try:
             self.forward(trace)
+            self.forward_afterburner(trace)
         except CoversTargetException:
             pass
         return trace
     
-    def online_resample(self, samples, dtype = torch.float32):
-        _offline = self.offline
-        self.offline = False
-        _online = self.online
-        self.online = True
-        out = []
-        for i in tqdm(range(len(samples))):
-            result = self._run(conditions = samples[i], targets = None, overwrite = True)
-            result = {k: self._to_tensor(v) for k, v in result.items()}
-            out.append(result)
-        self.offline = _offline
-        self.online = _online
-        return self._collate_output(out, dtype)
+    def apply_afterburner(self, samples, dtype = torch.float32):
+        trace = SwyftTrace(None, conditions = samples, overwrite = True)
+        self.forward_afterburner(trace)
+        return {k: self._to_tensor(v) for k, v in trace.items()}
     
     @staticmethod
     def _collate_output(out, dtype):
@@ -122,7 +115,7 @@ class SwyftModelForward:
             result = {k: self._to_tensor(v) for k, v in result.items()}
             out.append(result)
         return self._collate_output(out, dtype)
-    
+
 
 class SwyftModel:
     def _simulate(self, N, bounds = None, effective_prior = None):
@@ -168,17 +161,17 @@ def tensorboard_config(save_dir = "./lightning_logs", name = None, version = Non
 
 
 class SwyftDataModule(pl.LightningDataModule):
-    def __init__(self, model = None, store = None, batch_size: int = 32, validation_percentage = 0.2, manual_seed = None, train_multiply = 10 , num_workers = 0):
+    def __init__(self, simulator = None, store = None, batch_size: int = 32, validation_percentage = 0.2, manual_seed = None, train_multiply = 10 , num_workers = 0):
         super().__init__()
         self.store = store
-        self.model = model
+        self.simulator = simulator
         self.batch_size = batch_size
         self.num_workers = num_workers
         self.validation_percentage = validation_percentage
         self.train_multiply = train_multiply
 
     def setup(self, stage):
-        hook = None if self.model is None else self.model.noise        
+        hook = None if self.simulator is None else self.simulator.apply_afterburner
         self.dataset = _DictDataset(self.store, hook = hook)#, x_keys = ['data'], z_keys=['z'])
         n_train, n_valid = get_ntrain_nvalid(self.validation_percentage, len(self.dataset))
         self.dataset_train, self.dataset_valid = random_split(self.dataset, [n_train, n_valid], generator=torch.Generator().manual_seed(42))
