### 2022-04-01 #4205 @pdebuyl
Thank you @olgadoronina for the review!

@bkmi I will proceed to the editorial steps in the coming days, see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
