+++ swyft/lightning/components.py
@@ -1,3 +1,4 @@
+from abc import abstractmethod
 from dataclasses import dataclass, field
 from toolz.dicttoolz import valmap
 from typing import (
@@ -69,8 +70,12 @@ class SwyftModelForward:
     offline = True
     online = True
 
+    @abstractmethod
     def forward(self, trace):
         raise NotImplementedError
+
+    def forward_afterburner(self, trace):
+        return trace
 
     def _run(self, targets = None, conditions = {}, overwrite = False):
         trace = SwyftTrace(targets, conditions, overwrite = overwrite)
@@ -113,7 +118,7 @@ class SwyftModelForward:
     def _to_tensor(v):
         if isinstance(v, np.ndarray):
             v = torch.from_numpy(v)
-        return v
+        return v.cpu()
 
     def sample(self, N, targets = None, conditions = {}, dtype = torch.float32):
         out = []
@@ -190,9 +195,9 @@ class SwyftDataModule(pl.LightningDataModule):
     def val_dataloader(self):
         return torch.utils.data.DataLoader(self.dataset_valid, batch_size=self.batch_size, num_workers = self.num_workers)
     
-    # TODO: Deprecate
-    def predict_dataloader(self):
-        return torch.utils.data.DataLoader(self.dataset, batch_size=self.batch_size, num_workers = self.num_workers)
+    # # TODO: Deprecate
+    # def predict_dataloader(self):
+    #     return torch.utils.data.DataLoader(self.dataset, batch_size=self.batch_size, num_workers = self.num_workers)
     
     def test_dataloader(self):
         return torch.utils.data.DataLoader(self.dataset_test, batch_size=self.batch_size, num_workers = self.num_workers)
@@ -303,7 +308,7 @@ class SwyftModule(pl.LightningModule):
 
 class SwyftTrainer(pl.Trainer):
     def infer(self, model, dataloader, conditions = {}):
-        self.model._set_predict_conditions(conditions, {})
+        model._set_predict_conditions(conditions, {})
         ratio_batches = self.predict(model, dataloader)
         keys = ratio_batches[0].keys()
         d = {k: RatioSamples(
@@ -311,7 +316,7 @@ class SwyftTrainer(pl.Trainer):
                 torch.cat([r[k].ratios for r in ratio_batches])
                 ) for k in keys
             }
-        self.model._set_predict_conditions({}, {})  # Set it back to no conditioning
+        model._set_predict_conditions({}, {})  # Set it back to no conditioning
         return RatioSampleStore(**d)
     
 
