### 2022-04-01 cc55045
make SwyftModelForward an iterator, add SwyftOnlineDatastore
