### 2022-04-01 fb6b4f3
remove faulty torchstore exception
