### 2022-04-08 17a4fb8
Added ZarrStoreIterableDataset
