### 2022-04-09 a0e18a5
Add trace.record function, and rename SwyftModelForward -> SwyftSimulator
