### 2022-04-11 #4205 @pdebuyl
Minor comments

1. In the notebook `Examples - 2. Truncation` I had to re-run the cell

```py
store.add(n_training_samples + 0.01 * n_training_samples, prior, bound=bound)
store.simulate()
dataset = swyft.Dataset(n_training_samples, prior, store, bound = bound)
```

twice as else there would not be enough samples. I don't know if it is unexpected though as the sampling is not deterministic.

2. In the article, there is no section titled "Statement of Need". This is a mandatory part of the article according to JOSS' guidelines. The current section "Motivation" is a good fit for it though, you could simply rename it.

3. There is mention that the technique would not be convenient to include in existing ML packages. Would it be a possibility in the future, provided some changes are made to an hypothetical good package for inclusion? (PS: not pushing for it, but I think that it is interesting to let the readers -- possibly maintainers of such a package -- what prevents it from happening).
4. It is convenient that "swyft automates irksome matters like distributed computing and data storage". What does it mean in practice? Is it by using dask? In that case, it could be mentioned explicitly after the quoted sentence.
5. To download the example notebooks, I cloned the repo. It is probably not much of an issue for the developers of swyft but for prospective users a "download this notebook" link would be great (-> this is only a suggestion).

PS: in case of further updates to the article file, this does not require a new zenodo archive or update to the version.
