+++ swyft/lightning/components_noemi.py
@@ -1,0 +1,930 @@
+from abc import abstractmethod
+import math
+from dataclasses import dataclass, field
+from toolz.dicttoolz import valmap
+from typing import (
+    Callable,
+    Dict,
+    Hashable,
+    Optional,
+    Sequence,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
+import numpy as np
+import torch
+import torch.nn as nn
+from torch.nn import functional as F
+from torch.utils.data import random_split
+import pytorch_lightning as pl
+from tqdm import tqdm
+import swyft
+import swyft.utils
+from swyft.inference.marginalratioestimator import get_ntrain_nvalid
+import yaml
+import hydra
+
+import zarr
+import fasteners
+from pytorch_lightning import loggers as pl_loggers
+
+from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint
+from pytorch_lightning import loggers as pl_loggers
+from pytorch_lightning.callbacks.early_stopping import EarlyStopping
+import pytorch_lightning as pl
+
+from swyft.networks.standardization import OnlineStandardizingLayer
+
+
+#################################
+# Swyft lightning main components
+#################################
+
+class CoversTargetException(Exception):
+    pass
+
+class SwyftTrace(dict):
+    def __init__(self, targets = None, conditions = {}, overwrite = False):
+        super().__init__(conditions)
+        self._targets = targets
+        self._overwrite = overwrite
+
+    def contains_not(self, keys):
+        if isinstance(keys, str):
+            keys = [keys]
+        return not all([k in self.keys() for k in keys])
+        
+    @property
+    def covers_targets(self):
+        if self._targets is not None:
+            if all([k in self.keys() for k in self._targets]):
+                return True
+
+    def __setitem__(self, k, v):
+        if k not in self.keys() or self._overwrite:
+            super().__setitem__(k, v)
+        if self.covers_targets:
+            raise CoversTargetException
+
+
+class SwyftModelForward:
+    @abstractmethod
+    def forward(self, trace):
+        raise NotImplementedError
+
+    def forward_afterburner(self, trace):
+        return trace
+
+    def _run(self, targets = None, conditions = {}, overwrite = False):
+        trace = SwyftTrace(targets, conditions, overwrite = overwrite)
+        if trace.covers_targets:
+            return trace
+        try:
+            self.forward(trace)
+            self.forward_afterburner(trace)
+        except CoversTargetException:
+            pass
+        return trace
+    
+    def apply_afterburner(self, samples, dtype = torch.float32):
+        trace = SwyftTrace(None, conditions = samples, overwrite = True)
+        self.forward_afterburner(trace)
+        return {k: self._to_tensor(v) for k, v in trace.items()}
+    
+    @staticmethod
+    def _collate_output(out, dtype):
+        out = torch.utils.data.dataloader.default_collate(out)
+        if dtype:
+            for k, v in out.items():
+                out[k] = v.type(dtype)
+        return SampleStore(out)
+
+    def get_shapes(self):
+        sample = self()
+        shapes = {k: tuple(v.shape) for k, v in sample.items()}
+        return shapes
+
+    @staticmethod
+    def _to_tensor(v):
+        if isinstance(v, np.ndarray):
+            v = torch.from_numpy(v)
+        return v.cpu()
+
+    def sample(self, N, targets = None, conditions = {}, dtype = torch.float32):
+        out = []
+        for _ in tqdm(range(N)):
+            result = self._run(targets, conditions)
[…]
