+++ swyft/lightning/components.py
@@ -45,10 +45,9 @@ class CoversTargetException(Exception):
     pass
 
 class SwyftTrace(dict):
-    def __init__(self, targets = None, conditions = {}, overwrite = False):
+    def __init__(self, targets = None, conditions = {}):
         super().__init__(conditions)
         self._targets = targets
-        self._overwrite = overwrite
 
     # TODO: Deprecate
     def contains_not(self, keys):
@@ -69,13 +68,16 @@ class SwyftTrace(dict):
     # TODO: Deprecate
     def __setitem__(self, k, v):
         print("WARNING: To be deprecated. Use `record` instead.")
-        if k not in self.keys() or self._overwrite:
+        if k not in self.keys():
             super().__setitem__(k, v)
         if self.covers_targets:
             raise CoversTargetException
 
     def record(self, k, v, *args, **kwargs):
-        if k not in self.keys() or self._overwrite:
+        # TODO: Deprecate
+        if not callable(v):
+            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
+        if k not in self.keys():
             v = v(*args, **kwargs) if callable(v) else v
             super().__setitem__(k, v)
         if self.covers_targets:
@@ -83,6 +85,9 @@ class SwyftTrace(dict):
         return self[k]
 
     def lazy_record(self, k, v, *args, **kwargs):
+        # TODO: Deprecate
+        if not callable(v):
+            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
         fnc = lambda: self.record(k, v, *args, **kwargs)
         cache = [None]
         if self._targets is not None and k in self._targets:
@@ -94,8 +99,6 @@ class SwyftTrace(dict):
         return wrapper
 
 
-
-
 class SwyftSimulator:
     @abstractmethod
     def forward(self, trace):
@@ -104,30 +107,42 @@ class SwyftSimulator:
     def forward_afterburner(self, trace):
         return trace
 
-    def _run(self, targets = None, conditions = {}, overwrite = False):
+    def __call__(self, targets = None, conditions = {}):
         conditions = conditions() if callable(conditions) else conditions
-        trace = SwyftTrace(targets, conditions, overwrite = overwrite)
-        if trace.covers_targets:
-            return trace
-        try:
-            self.forward(trace)
-            self.forward_afterburner(trace)
-        except CoversTargetException:
-            pass
-        return trace
+        trace = SwyftTrace(targets, conditions)
+        if not trace.covers_targets:
+            try:
+                self.forward(trace)
+                #self.forward_afterburner(trace)
+            except CoversTargetException:
+                pass
+        result = {k: self._to_tensor(v) for k, v in trace.items()}
+        return result
     
     def apply_afterburner(self, samples, dtype = torch.float32):
-        trace = SwyftTrace(None, conditions = samples, overwrite = True)
+        print("Warning: To be deprecated.")
+        trace = SwyftTrace(None, conditions = samples)
         self.forward_afterburner(trace)
         return {k: self._to_tensor(v) for k, v in trace.items()}
-    
+#    
+#    @staticmethod
+#    def _collate_output(out, dtype):
+#        out = torch.utils.data.dataloader.default_collate(out)
+#        #if dtype:
+#        #    for k, v in out.items():
+#        #        out[k] = v.type(dtype)
+#        return SampleStore(out)
+
     @staticmethod
-    def _collate_output(out, dtype):
-        out = torch.utils.data.dataloader.default_collate(out)
-        if dtype:
-            for k, v in out.items():
-                out[k] = v.type(dtype)
-        return SampleStore(out)
+    def _collate_output(out):
[…]
