+++ swyft/lightning/components.py
@@ -49,6 +49,9 @@ class SwyftTrace(dict):
         super().__init__(conditions)
         self._targets = targets
 
+    def __repr__(self):
+        return "SwyftTrace("+super().__repr__()+")"
+
     # TODO: Deprecate
     def contains_not(self, keys):
         print("WARNING: To be deprecated. Use `contains` instead.")
@@ -100,30 +103,36 @@ class SwyftTrace(dict):
 
 
 class SwyftSimulator:
+    def on_before_forward(self, sample):
+        return sample
+
     @abstractmethod
     def forward(self, trace):
         raise NotImplementedError
 
-    def forward_afterburner(self, trace):
-        return trace
+    def on_after_forward(self, sample):
+        return sample
 
     def __call__(self, targets = None, conditions = {}):
+        # Evaluate conditions if callable
         conditions = conditions() if callable(conditions) else conditions
+
+        conditions = self.on_before_forward(conditions)
         trace = SwyftTrace(targets, conditions)
         if not trace.covers_targets:
             try:
                 self.forward(trace)
-                #self.forward_afterburner(trace)
             except CoversTargetException:
                 pass
-        result = {k: self._to_tensor(v) for k, v in trace.items()}
+        result = self.on_after_forward(dict(trace))
+
         return result
     
-    def apply_afterburner(self, samples, dtype = torch.float32):
-        print("Warning: To be deprecated.")
-        trace = SwyftTrace(None, conditions = samples)
-        self.forward_afterburner(trace)
-        return {k: self._to_tensor(v) for k, v in trace.items()}
+#    def apply_afterburner(self, samples, dtype = torch.float32):
+#        print("Warning: To be deprecated.")
+#        trace = SwyftTrace(None, conditions = samples)
+#        self.forward_afterburner(trace)
+#        return {k: self._to_tensor(v) for k, v in trace.items()}
 #    
 #    @staticmethod
 #    def _collate_output(out, dtype):
@@ -144,17 +153,18 @@ class SwyftSimulator:
                 result[key] = np.stack([x[key] for x in out])
         return result
 
-    def get_shapes(self):
-        sample = self()
+    def get_shapes_and_dtypes(self, targets = None):
+        sample = self(targets = targets)
         shapes = {k: tuple(v.shape) for k, v in sample.items()}
-        return shapes
+        dtypes = {k: v.dtype for k, v in sample.items()}
+        return shapes, dtypes
 
-    @staticmethod
-    def _to_tensor(v):
-        return v
-        if isinstance(v, np.ndarray):
-            v = torch.from_numpy(v)
-        return v.cpu()
+#    @staticmethod
+#    def _to_tensor(v):
+#        return v
+#        if isinstance(v, np.ndarray):
+#            v = torch.from_numpy(v)
+#        return v.cpu()
 
     def sample(self, N, targets = None, conditions = {}):
         out = []
@@ -806,27 +816,16 @@ class ZarrStore:
         self.lock = fasteners.InterProcessLock(file_path+".lock.file")
             
     def reset_sim_status(self):
-        pass
+        raise NotImplementedError
     
     def extend(self, N):
-        pass
+        raise NotImplementedError
         
-    @staticmethod
-    def _get_shapes(model, shapes):
-        if model is None and shapes is not None:
-            pass
-        elif model is not None and shapes is None:
-            sample = model.sample(1)[0]
-            shapes = {k: tuple(v.shape) for k, v in sample.items()}
-        else:
-            raise ValueError("Either shapes or model must be specified (but not both).")
-        return shapes
-    
[…]
