+++ swyft/lightning/components_v2.py
@@ -1,1053 +1,0 @@
-from abc import abstractmethod
-import math
-from dataclasses import dataclass, field
-from toolz.dicttoolz import valmap
-from typing import (
-    Callable,
-    Dict,
-    Hashable,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-)
-import numpy as np
-import torch
-import torch.nn as nn
-from torch.nn import functional as F
-from torch.utils.data import random_split
-import pytorch_lightning as pl
-from tqdm import tqdm
-import swyft
-import swyft.utils
-from swyft.inference.marginalratioestimator import get_ntrain_nvalid
-import yaml
-
-import zarr
-import fasteners
-from pytorch_lightning import loggers as pl_loggers
-
-from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint
-from pytorch_lightning import loggers as pl_loggers
-from pytorch_lightning.callbacks.early_stopping import EarlyStopping
-import pytorch_lightning as pl
-
-from swyft.networks.standardization import OnlineStandardizingLayer
-
-
-#################################
-# Swyft lightning main components
-#################################
-
-class CoversTargetException(Exception):
-    pass
-
-class SwyftTrace(dict):
-    def __init__(self, targets = None, conditions = {}):
-        super().__init__(conditions)
-        self._targets = targets
-
-    def __repr__(self):
-        return "SwyftTrace("+super().__repr__()+")"
-
-    # TODO: Deprecate
-    def contains_not(self, keys):
-        print("WARNING: To be deprecated. Use `contains` instead.")
-        return not self.contains(keys)
-    
-    def contains(self, keys):
-        if isinstance(keys, str):
-            keys = [keys]
-        return all([k in self.keys() for k in keys])
-        
-    @property
-    def covers_targets(self):
-        if self._targets is not None:
-            if all([k in self.keys() for k in self._targets]):
-                return True
-
-    # TODO: Deprecate
-    def __setitem__(self, k, v):
-        print("WARNING: To be deprecated. Use `record` instead.")
-        if k not in self.keys():
-            super().__setitem__(k, v)
-        if self.covers_targets:
-            raise CoversTargetException
-
-    def record(self, k, v, *args, **kwargs):
-        # TODO: Deprecate
-        if not callable(v):
-            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
-        if k not in self.keys():
-            v = v(*args, **kwargs) if callable(v) else v
-            super().__setitem__(k, v)
-        if self.covers_targets:
-            raise CoversTargetException
-        return self[k]
-
-    def lazy_record(self, k, v, *args, **kwargs):
-        # TODO: Deprecate
-        if not callable(v):
-            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
-        fnc = lambda: self.record(k, v, *args, **kwargs)
-        cache = [None]
-        if self._targets is not None and k in self._targets:
-            cache[0] = fnc()
-        def wrapper(cache = cache):
-            if cache[0] is None:
-                cache[0] = fnc()
-            return cache[0]
-        return wrapper
-
-
-class SwyftSimulator:
-    def on_before_forward(self, sample):
-        return sample
-
-    @abstractmethod
-    def forward(self, trace):
-        raise NotImplementedError
-
-    def on_after_forward(self, sample):
-        return sample
-
-    def __call__(self, targets = None, conditions = {}):
-        # Evaluate conditions if callable
[…]
