+++ swyft/lightning/components.py
@@ -45,15 +45,22 @@ class CoversTargetException(Exception):
     pass
 
 class SwyftTrace(dict):
-    def __init__(self, targets = None, conditions = {}, overwrite = False):
+    def __init__(self, targets = None, conditions = {}):
         super().__init__(conditions)
         self._targets = targets
-        self._overwrite = overwrite
 
+    def __repr__(self):
+        return "SwyftTrace("+super().__repr__()+")"
+
+    # TODO: Deprecate
     def contains_not(self, keys):
+        print("WARNING: To be deprecated. Use `contains` instead.")
+        return not self.contains(keys)
+    
+    def contains(self, keys):
         if isinstance(keys, str):
             keys = [keys]
-        return not all([k in self.keys() for k in keys])
+        return all([k in self.keys() for k in keys])
         
     @property
     def covers_targets(self):
@@ -63,78 +70,139 @@ class SwyftTrace(dict):
 
     # TODO: Deprecate
     def __setitem__(self, k, v):
-        if k not in self.keys() or self._overwrite:
+        print("WARNING: To be deprecated. Use `record` instead.")
+        if k not in self.keys():
             super().__setitem__(k, v)
         if self.covers_targets:
             raise CoversTargetException
 
-    def record(self, k, v):
-        if k not in self.keys() or self._overwrite:
-            v = v() if callable(v) else v
+    def record(self, k, v, *args, **kwargs):
+        # TODO: Deprecate
+        if not callable(v):
+            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
+        if k not in self.keys():
+            v = v(*args, **kwargs) if callable(v) else v
             super().__setitem__(k, v)
         if self.covers_targets:
             raise CoversTargetException
         return self[k]
 
+    def lazy_record(self, k, v, *args, **kwargs):
+        # TODO: Deprecate
+        if not callable(v):
+            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
+        fnc = lambda: self.record(k, v, *args, **kwargs)
+        cache = [None]
+        if self._targets is not None and k in self._targets:
+            cache[0] = fnc()
+        def wrapper(cache = cache):
+            if cache[0] is None:
+                cache[0] = fnc()
+            return cache[0]
+        return wrapper
 
 
 class SwyftSimulator:
+    def on_before_forward(self, sample):
+        return sample
+
     @abstractmethod
     def forward(self, trace):
         raise NotImplementedError
 
-    def forward_afterburner(self, trace):
-        return trace
+    def on_after_forward(self, sample):
+        return sample
 
-    def _run(self, targets = None, conditions = {}, overwrite = False):
-        trace = SwyftTrace(targets, conditions, overwrite = overwrite)
-        if trace.covers_targets:
-            return trace
-        try:
-            self.forward(trace)
-            self.forward_afterburner(trace)
-        except CoversTargetException:
-            pass
-        return trace
+    def __call__(self, targets = None, conditions = {}):
+        # Evaluate conditions if callable
+        conditions = conditions() if callable(conditions) else conditions
+
+        conditions = self.on_before_forward(conditions)
+        trace = SwyftTrace(targets, conditions)
+        if not trace.covers_targets:
+            try:
+                self.forward(trace)
+            except CoversTargetException:
+                pass
+        result = self.on_after_forward(dict(trace))
+
+        return result
     
-    def apply_afterburner(self, samples, dtype = torch.float32):
[…]
