+++ swyft/lightning/components.py
@@ -37,13 +37,6 @@ import pytorch_lightning as pl
 from swyft.networks.standardization import OnlineStandardizingLayer
 
 
-#################################
-# Swyft lightning main components
-#################################
-
-class CoversTargetException(Exception):
-    pass
-
 class SwyftTrace(dict):
     def __init__(self, targets = None, conditions = {}):
         super().__init__(conditions)
@@ -52,54 +45,48 @@ class SwyftTrace(dict):
     def __repr__(self):
         return "SwyftTrace("+super().__repr__()+")"
 
-    # TODO: Deprecate
-    def contains_not(self, keys):
-        print("WARNING: To be deprecated. Use `contains` instead.")
-        return not self.contains(keys)
-    
-    def contains(self, keys):
-        if isinstance(keys, str):
-            keys = [keys]
-        return all([k in self.keys() for k in keys])
-        
     @property
     def covers_targets(self):
-        if self._targets is not None:
-            if all([k in self.keys() for k in self._targets]):
-                return True
+        return (self._targets is not None 
+                and all([k in self.keys() for k in self._targets]))
 
-    # TODO: Deprecate
-    def __setitem__(self, k, v):
-        print("WARNING: To be deprecated. Use `record` instead.")
-        if k not in self.keys():
-            super().__setitem__(k, v)
-        if self.covers_targets:
-            raise CoversTargetException
+    def sample(self, name, fn, *args, **kwargs):
+        assert callable(fn), "Second argument must be a function."
+        lazy_value = SwyftLazyValue(self, name, fn, *args, **kwargs)
+        if self._targets is None or name in self._targets:
+            lazy_value.evaluate()
+        return lazy_value
 
-    def record(self, k, v, *args, **kwargs):
-        # TODO: Deprecate
-        if not callable(v):
-            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
-        if k not in self.keys():
-            v = v(*args, **kwargs) if callable(v) else v
-            super().__setitem__(k, v)
-        if self.covers_targets:
-            raise CoversTargetException
-        return self[k]
 
-    def lazy_record(self, k, v, *args, **kwargs):
-        # TODO: Deprecate
-        if not callable(v):
-            print("WARNING: Second argument should be a function.  Variables deprecated soon.")
-        fnc = lambda: self.record(k, v, *args, **kwargs)
-        cache = [None]
-        if self._targets is not None and k in self._targets:
-            cache[0] = fnc()
-        def wrapper(cache = cache):
-            if cache[0] is None:
-                cache[0] = fnc()
-            return cache[0]
-        return wrapper
+class SwyftLazyValue:
+    def __init__(self, trace, name, fn, *args, **kwargs):
+        self._trace = trace
+        self._name = name
+        self._fn = fn
+        self._args = args
+        self._kwargs = kwargs
+
+    def __repr__(self):
+        value = self._trace[self._name] if self._name in self._trace.keys() else "None"
+        return f"SwyftLazyValue{self._name, value, self._fn, self._args, self._kwargs}"
+
+    def evaluate(self):
+        if self._name not in self._trace.keys():
+            args = (arg.evaluate() if isinstance(arg, SwyftLazyValue) else arg for arg in self._args)
+            kwargs = {k: v.evaluate() if isinstance(v, SwyftLazyValue) else v for k, v in self._kwargs.items()}
+            self._trace[self._name] = self._fn(*args, **kwargs)
+        return self._trace[self._name]
+
+
+def collate_output(out):
+    keys = out[0].keys()
+    result = {}
+    for key in keys:
[…]
