+++ paper/paper.md
@@ -47,10 +47,10 @@ Parametric stochastic numerical simulators are ubiquitous in science. They model
 Our package `swyft` implements a specific, simulation-efficient SBI method called *Truncated Marginal Neural Ratio Estimation* (TMNRE) [@miller2021truncated]; it estimates the likelihood-to-evidence ratio to approximate the posterior, as in @Hermans2019. `swyft` [@swyft] provides a collection of tools to simulate and store data, locally or in a distributed computing setting, and perform (marginalized) simulation-based Bayesian inference. It produces ready-to-publish plots that demonstrate the calibration of the posterior estimate along with the posterior itself.
 
 
-# Motivation
+# Statement of Need
 Estimating the posterior can be prohibitively expensive for complex data and slow simulators. Part of the reason is the sequential nature of likelihood-based Markov chain Monte-Carlo [@metropolis; @hastings]. In contrast, SBI parallelizes simulation in most circumstances, thereby reducing the practical waiting time for results. In pursuit of further simulation efficiency, @miller2021truncated argue that fitting the joint posterior for all parameters is unnecessary when a marginal estimate of the posterior will suffice. Some SBI methods are amortized, whereby the statistical model is fit to estimate posteriors for all possible observations simultaneously. While amortization enables necessary posterior calibration checks, like expected coverage probability [@miller2021truncated; @hermans2021averting], it is more efficient to fit the model on only a subset of the parameters that could have plausibly generated the observation.
 
-`swyft` satisfies necessary requirements, like estimating the marginal posteriors of interest and enabling posterior calibration checks, while taking a lean approach to avoid all unnecessary simulation. In this pursuit, `swyft` truncates the prior to regions relevant for given observational data and reuses compatible existing simulations.  `swyft` automates irksome matters like distributed computing and data storage. `swyft` is designed to:
+`swyft` satisfies necessary requirements, like estimating the marginal posteriors of interest and enabling posterior calibration checks, while taking a lean approach to avoid all unnecessary simulation. In this pursuit, `swyft` truncates the prior to regions relevant for given observational data and reuses compatible existing simulations.  `swyft` automates irksome matters like distributed computing and data storage with `dask` [@dask] and `zarr` [@zarr] respectively. `swyft` is designed to:
 
 1. Estimate arbitrary marginal posteriors, i.e., the posterior over parameters of interest, marginalizing over nuisance parameters.
 2. Perform targeted inference by truncating the prior distribution with an indicator function estimated in a sequence of inferences.
@@ -58,7 +58,7 @@ Estimating the posterior can be prohibitively expensive for complex data and slo
 4. Seamlessly reuse simulations from previous analyses by drawing already-simulated data first via a flexible storage solution.
 5. Integrate advanced distribution and storage tools to simplify application of complex simulators.
 
-Although there is a rich ecosystem of SBI implementations, TMNRE did not naturally fit in an existing framework since it requires parallel estimation of marginal posteriors and a truncated prior. `swyft` aims to meet the ever-increasing demand for efficient and testable Bayesian inference in fields like physics, cosmology, and astronomy by implementing TMNRE together with practical distributed computing and storage tools.
+Although there is a rich ecosystem of SBI implementations, TMNRE did not naturally fit in an existing framework since it requires parallel estimation of marginal posteriors and a truncated prior. `swyft` does the parallel training of the ratio using another dimension in a PyTorch tensor and created a custom truncated prior data structure to overcome these challenges. `swyft` aims to meet the ever-increasing demand for efficient and testable Bayesian inference in fields like physics, cosmology, and astronomy by implementing TMNRE together with practical distributed computing and storage tools.
 
 ## Existing research with `swyft`
[…]
