### 2022-06-10 7fbde23
updated docs to include link to notebooks
