+++ docs/source/conf.py
@@ -20,6 +20,8 @@
 # import sys
 # sys.path.insert(0, os.path.abspath("../../"))
 
+import os
+
 import sphinx_rtd_theme
 from pkg_resources import get_distribution
 
@@ -34,7 +36,38 @@ from pkg_resources import get_distribution
 # ]
 
 
+# -- Custom project information --------------------------------------------------------
+
+here = os.path.dirname(__file__)
+repo = os.path.join(here, "..", "..")
+# _version_py = os.path.join(repo, 'swyft', '__version__.py')
+# version_ns = {}
+# with open(_version_py) as f:
+#     exec(f.read(), version_ns)
+# # The short X.Y version.
+# version = '%i.%i' % version_ns['version_info'][:2]
+# # The full version, including alpha/beta/rc tags.
+# release = version_ns['__version__']
+
+import subprocess
+
+try:
+    git_rev = subprocess.check_output(
+        ["git", "describe", "--exact-match", "HEAD"], universal_newlines=True
+    )
+except subprocess.CalledProcessError:
+    try:
+        git_rev = subprocess.check_output(
+            ["git", "rev-parse", "HEAD"], universal_newlines=True
+        )
+    except subprocess.CalledProcessError:
+        git_rev = ""
+if git_rev:
+    git_rev = git_rev.splitlines()[0] + "/"
+
+
 # -- Custom config --------------------------------------------------------
+
 
 autodoc_typehints = "description"
 napoleon_numpy_docstring = False
@@ -67,17 +100,58 @@ nbsphinx_allow_errors = True
 # autosummary_generate = True
 
 # This is processed by Jinja2 and inserted before each notebook
-# idea taken from: https://github.com/stellargraph/stellargraph/pull/1460/files
-nbsphinx_prolog = r"""
-.. raw:: html
-    <div class="admonition info">
-        <p>
-        <a href="{{ env.docname.rsplit('/', 1).pop() }}.ipynb" class="btn">Download this notebook</a></p>
-        </p>
-    </div>
-"""
+# idea taken from: https://github.com/vidartf/nbsphinx-link/blob/f1682a8b5bfd884bfed7ef7dfaf0810c6ae8a543/docs/source/conf.py
 
-nbsphinx_epilog = nbsphinx_prolog  # also insert after each notebook
+nbsphinx_link_target_root = repo
+
+nbsphinx_prolog = (
+    r"""
+{% if env.metadata[env.docname]['nbsphinx-link-target'] %}
+{% set docpath = env.metadata[env.docname]['nbsphinx-link-target'] %}
+{% else %}
+{% set docpath = env.doc2path(env.docname, base='docs/source/') %}
+{% endif %}
+
+.. only:: html
+
+    .. role:: raw-html(raw)
+        :format: html
+
+    .. nbinfo::
+        This page was generated from `{{ docpath }}`_.
+        {{ docpath.split()[0] }}
+        {{ '\ '.join(docpath.split()) }}
+
+    .. _{{ docpath }}: https://github.com/undark-lab/swyft/blob/ """
+    + git_rev
+    + r"{{ '\ '.join(docpath.split()) }}"  # it is required to join and split because spaces must be escaped in rst!!
+)
+
+# nbsphinx_epilog = nbsphinx_prolog  # also insert after each notebook
+
+
+# nbsphinx_prolog = r"""
+# {% set docname = env.doc2path(env.docname, base=None) %}
+
+# .. only:: html
+
+#     Go there: https://example.org/notebooks/{{ docname }}
+
+# .. raw:: latex
+
+#     \nbsphinxstartnotebook{The following section was created from
+#     \texttt{\strut{}{{ docname }}}:}
+# """
+
+
[…]
