### 2022-06-10 a9aad7e
updated to remove extra info in prolog
