+++ docs/source/conf.py
@@ -125,31 +125,7 @@ nbsphinx_prolog = (
     + r"{{ '\ '.join(docpath.split()) }}"  # it is required to join and split because spaces must be escaped in rst!!
 )
 
-# nbsphinx_epilog = nbsphinx_prolog  # also insert after each notebook
-
-
-# nbsphinx_prolog = r"""
-# {% set docname = env.doc2path(env.docname, base=None) %}
-
-# .. only:: html
-
-#     Go there: https://example.org/notebooks/{{ docname }}
-
-# .. raw:: latex
-
-#     \nbsphinxstartnotebook{The following section was created from
-#     \texttt{\strut{}{{ docname }}}:}
-# """
-
-
-nbsphinx_epilog = """
-----
-
-Generated by nbsphinx_ from a Jupyter_ notebook.
-
-.. _nbsphinx: https://nbsphinx.readthedocs.io/
-.. _Jupyter: https://jupyter.org/
-"""
+nbsphinx_epilog = nbsphinx_prolog  # also insert after each notebook
 
 
 # -- General configuration ------------------------------------------------
