### 2022-06-13 #4205 @pdebuyl
Hi @bkmi thanks for the update. It is in good shape :-) The changes do not impact the code, only the documentation configuration, so that there is no need for a further release.

Can you edit the metadata of the zenodo entry so that it matches the one of the paper?
- Title should be "swyft: Truncated Marginal Neural Ratio Estimation in Python"
- Authors should be the same and in the same order. If someone is a contributor but not an author of the paper, you can move them to "contributor" on zenodo instead of author.
