### 2022-06-19 #4205 @bkmi
The author order is not allowed to be altered? The reason we want to is because the largest code contributor is the PI of the lab. Also, a significant amount of the code, written by some contributors, is ipython notebook output.

Am I understanding it right that it's possible to move people off the author list who have contributed code as long as they're are added to the contributors list?

The other aspects are no problem.
