### 2022-06-19 #4205 @pdebuyl
> The author order is not allowed to be altered? The reason we want to is because the largest code contributor is the PI of the lab. Also, a significant amount of the code, written by some contributors, is ipython notebook output.

From Zenodo's FAQ:

> I only want to change the title of my upload, do I still get a new DOI?
>
> No, as before you can continue to edit the metadata of your upload without creating a new version of a record. You should only create a new version if you want to update the files of your record.

So you can change Zenodo's author order and update the record. You cannot change the files in a record.

JOSS policy is that the author list between the paper and archive must match. See in the editorial guide https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance I can't find the same guideline in the "author info" part of our documentation. If this is an issue for you, I'll query the editorial team.

> Am I understanding it right that it's possible to move people off the author list who have contributed code as long as they're are added to the contributors list?

Yes. For many JOSS articles, there is a group of main contributors that are authors of the article while some contributors are not authors. See a relevant discussion here for instance: https://github.com/openjournals/joss/issues/815
