### 2022-06-20 dfaa33d
Updates loss function
