### 2022-06-20 1f8f0b7
Wrote better predict and infer routines
