+++ swyft/lightning/components.py
@@ -39,14 +39,23 @@ from pytorch_lightning.trainer.supporters import CombinedLoader
 from swyft.networks.standardization import OnlineStandardizingLayer
 
 
-class SwyftTrace(dict):
+class Trace(dict):
+    """Defines the computational graph (DAG) and keeps track of simulation results.
+    """
     def __init__(self, targets = None, conditions = {}):
+        """Instantiate Trace instante.
+
+        Args:
+            targets: Optional list of target sample variables. If provided, execution is stopped after those targets are evaluated. If `None`, all variables in the DAG will be evaluated.
+            conditions: Optional `dict` or Callable. If a `dict`, sample variables will be conditioned to the corresponding values.  If Callable, it will be evaulated and it is expected to return a `dict`.
+        """
+
         super().__init__(conditions)
         self._targets = targets
         self._prefix = ""
 
     def __repr__(self):
-        return "SwyftTrace("+super().__repr__()+")"
+        return "Trace("+super().__repr__()+")"
 
     def __setitem__(self, k, v):
         if k not in self.keys():
@@ -58,6 +67,16 @@ class SwyftTrace(dict):
                 and all([k in self.keys() for k in self._targets]))
 
     def sample(self, names, fn, *args, **kwargs):
+        """Register sampling function.
+
+        Args:
+            names: Name or list of names of sampling variables.
+            fn: Callable that returns the (list of) sampling variable(s).
+            *args, **kwargs: Arguments and keywords arguments that are passed to `fn` upon evaluation.  LazyValues will be automatically evaluated if necessary.
+
+        Returns:
+            LazyValue sample.
+        """
         assert callable(fn), "Second argument must be a function."
         if isinstance(names, list):
             names = [self._prefix + n for n in names]
@@ -91,7 +110,18 @@ class TracePrefixContextManager:
 
 
 class LazyValue:
+    """Provides lazy evaluation functionality.
+    """
     def __init__(self, trace, this_name, fn_out_names, fn, *args, **kwargs):
+        """Instantiates LazyValue object.
+
+        Args:
+            trace: Trace instance (to be populated with sample).
+            this_name: Name of this the variable that this LazyValue represents.
+            fn_out_names: Name or list of names of variables that `fn` returns.
+            fn: Callable that returns sample or list of samples.
+            args, kwargs: Arguments and keyword arguments provided to `fn` upon evaluation.
+        """
         self._trace = trace
         self._this_name = this_name
         self._fn_out_names = fn_out_names
@@ -105,9 +135,15 @@ class LazyValue:
 
     @property
     def value(self):
+        """Value of this object."""
         return self.evaluate()
 
     def evaluate(self):
+        """Trigger evaluation of function.
+
+        Returns:
+            Value of `this_name`.
+        """
         if self._this_name not in self._trace.keys():
             args = (arg.evaluate() if isinstance(arg, LazyValue) else arg for arg in self._args)
             kwargs = {k: v.evaluate() if isinstance(v, LazyValue) else v for k, v in self._kwargs.items()}
@@ -133,21 +169,26 @@ def collate_output(out):
 
 
 class SwyftSimulator:
+    """Handles simulations."""
     def on_before_forward(self, sample):
+        """Apply transformations to conditions."""
         return sample
 
     @abstractmethod
     def forward(self, trace):
+        """Main function to overwrite.
+        """
         raise NotImplementedError
 
     def on_after_forward(self, sample):
+        """Apply transformation to generated samples."""
         return sample
 
     def __call__(self, targets = None, conditions = {}):
[…]
