+++ swyft/lightning/components.py
@@ -39,6 +39,47 @@ from pytorch_lightning.trainer.supporters import CombinedLoader
 from swyft.networks.standardization import OnlineStandardizingLayer
 
 
+################
+# SwyftSimulator
+################
+
+def to_numpy(x, single_precision = False):
+    if isinstance(x, torch.Tensor):
+        if not single_precision:
+            return x.detach().cpu().numpy()
+        else:
+            x = x.detach().cpu()
+            if x.dtype == torch.float64:
+                x = x.float().numpy()
+            else:
+                x = x.numpy()
+            return x
+    elif isinstance(x, SampleStore):
+        return SampleStore({k: to_numpy(v, single_precision = single_precision) for k, v in x.items()})
+    elif isinstance(x, dict):
+        return {k: to_numpy(v, single_precision = single_precision) for k, v in x.items()}
+    elif isinstance(x, np.ndarray):
+        if not single_precision:
+            return x
+        else:
+            if x.dtype == np.float64:
+                x = np.float32(x)
+            return x
+    else:
+        return x
+
+def to_numpy32(x):
+    return to_numpy(x, single_precision = True)
+    
+def to_torch(x):
+    if isinstance(x, SampleStore):
+        return SampleStore({k: to_torch(v) for k, v in x.items()})
+    elif isinstance(x, dict):
+        return {k: to_torch(v) for k, v in x.items()}
+    else:
+        return torch.as_tensor(x)
+    
+
 class Trace(dict):
     """Defines the computational graph (DAG) and keeps track of simulation results.
     """
@@ -158,6 +199,7 @@ class LazyValue:
 
 
 def collate_output(out):
+    """Turn list of tensors/arrays-value dicts into dict of collated tensors or arrays"""
     keys = out[0].keys()
     result = {}
     for key in keys:
@@ -282,86 +324,11 @@ class SwyftSimulatorResampler:
             conditions.pop(k)
         sims = self._simulator(conditions = conditions, targets = self._targets)
         return sims
-    
-
-def tensorboard_config(save_dir = "./lightning_logs", name = None, version = None, patience = 3):
-    """Generates convenience configuration for Trainer object.
-
-    Args:
-        save_dir: Save-directory for tensorboard logs
-        name: tensorboard logs name
-        version: tensorboard logs version
-        patience: early-stopping patience
-
-    Returns:
-        Configuration dictionary
-    """
-    tbl = pl_loggers.TensorBoardLogger(save_dir = save_dir, name = name, version = version, default_hp_metric = False)
-    lr_monitor = LearningRateMonitor(logging_interval="step")
-    early_stop_callback = EarlyStopping(monitor="val_loss", min_delta=0.0, patience=patience, verbose=False, mode="min")
-    checkpoint_callback = ModelCheckpoint(monitor="val_loss")
-    return dict(logger = tbl, callbacks = [lr_monitor, early_stop_callback, checkpoint_callback])
 
 
-# TODO: DEPRECATE?
-class SwyftDataModule(pl.LightningDataModule):
-    def __init__(self, on_after_load_sample = None, store = None, batch_size: int = 32, validation_percentage = 0.2, manual_seed = None, train_multiply = 10 , num_workers = 0):
-        super().__init__()
-        self.store = store
-        self.on_after_load_sample = on_after_load_sample
-        self.batch_size = batch_size
-        self.num_workers = num_workers
-        self.validation_percentage = validation_percentage
-        self.train_multiply = train_multiply
[…]
