+++ swyft/lightning/components.py
@@ -39,9 +39,9 @@ from pytorch_lightning.trainer.supporters import CombinedLoader
 from swyft.networks.standardization import OnlineStandardizingLayer
 
 
-################
-# SwyftSimulator
-################
+###########
+# Simulator
+###########
 
 def to_numpy(x, single_precision = False):
     if isinstance(x, torch.Tensor):
@@ -54,8 +54,8 @@ def to_numpy(x, single_precision = False):
             else:
                 x = x.numpy()
             return x
-    elif isinstance(x, SampleStore):
-        return SampleStore({k: to_numpy(v, single_precision = single_precision) for k, v in x.items()})
+    elif isinstance(x, Samples):
+        return Samples({k: to_numpy(v, single_precision = single_precision) for k, v in x.items()})
     elif isinstance(x, dict):
         return {k: to_numpy(v, single_precision = single_precision) for k, v in x.items()}
     elif isinstance(x, np.ndarray):
@@ -72,8 +72,8 @@ def to_numpy32(x):
     return to_numpy(x, single_precision = True)
     
 def to_torch(x):
-    if isinstance(x, SampleStore):
-        return SampleStore({k: to_torch(v) for k, v in x.items()})
+    if isinstance(x, Samples):
+        return Samples({k: to_torch(v) for k, v in x.items()})
     elif isinstance(x, dict):
         return {k: to_torch(v) for k, v in x.items()}
     else:
@@ -210,7 +210,7 @@ def collate_output(out):
     return result
 
 
-class SwyftSimulator:
+class Simulator:
     """Handles simulations."""
     def on_before_forward(self, sample):
         """Apply transformations to conditions."""
@@ -270,7 +270,7 @@ class SwyftSimulator:
             result = self.__call__(targets, conditions)
             out.append(result)
         out = collate_output(out)
-        out = SampleStore(out)
+        out = Samples(out)
         return out
 
     def get_resampler(self, targets):
@@ -280,9 +280,9 @@ class SwyftSimulator:
             targets: List of target variables to simulate
 
         Returns:
-            SwyftSimulatorResampler instance.
+            SimulatorResampler instance.
         """
-        return SwyftSimulatorResampler(self, targets)
+        return SimulatorResampler(self, targets)
 
     def get_iterator(self, targets = None, conditions = {}):
         """Generates an iterator. Useful for iterative sampling.
@@ -298,10 +298,10 @@ class SwyftSimulator:
         return iterator
     
     
-class SwyftSimulatorResampler:
+class SimulatorResampler:
     """Handles rerunning part of the simulator. Typically used for on-the-fly calculations during training."""
     def __init__(self, simulator, targets):
-        """Instantiates SwyftSimulatorResampler
+        """Instantiates SimulatorResampler
 
         Args:
             simulator: The simulator object
@@ -446,22 +446,22 @@ class SwyftTrainer(pl.Trainer):
             Concatenated network output
         """
         if isinstance(A, dict):
-            dl1 = SampleStore({k: [v] for k, v in A.items()}).get_dataloader(batch_size = 1)
+            dl1 = Samples({k: [v] for k, v in A.items()}).get_dataloader(batch_size = 1)
         else:
             dl1 = A
         if isinstance(B, dict):
-            dl2 = SampleStore({k: [v] for k, v in B.items()}).get_dataloader(batch_size = 1)
+            dl2 = Samples({k: [v] for k, v in B.items()}).get_dataloader(batch_size = 1)
         else:
             dl2 = B
         dl = CombinedLoader([dl1, dl2], mode = 'max_size_cycle')
[…]
