### 2022-07-01 c175075
Add log_prob arguments
