### 2022-07-02 e95b4cc
Revert introduction of log_prob components
