+++ swyft/lightning/components.py
@@ -84,19 +84,17 @@ def to_torch(x):
 class Trace(dict):
     """Defines the computational graph (DAG) and keeps track of simulation results.
     """
-    def __init__(self, targets = None, conditions = {}, log_prob = False):
+    def __init__(self, targets = None, conditions = {}):
         """Instantiate Trace instante.
 
         Args:
             targets: Optional list of target sample variables. If provided, execution is stopped after those targets are evaluated. If `None`, all variables in the DAG will be evaluated.
             conditions: Optional `dict` or Callable. If a `dict`, sample variables will be conditioned to the corresponding values.  If Callable, it will be evaulated and it is expected to return a `dict`.
-            log_prob: Boolean. Collect log_prob values if available. Default False.
         """
 
         super().__init__(conditions)
         self._targets = targets
         self._prefix = ""
-        self._eval_log_probs = log_prob
 
     def __repr__(self):
         return "Trace("+super().__repr__()+")"
@@ -122,21 +120,15 @@ class Trace(dict):
             LazyValue sample.
         """
         assert callable(fn), "Second argument must be a function."
-        return self._sample(names, fn, False, *args, **kwargs)
-
-    def sample_dist(self, names, fn, *args, **kwargs):
-        return self._sample(names, fn, True, *args, **kwargs)
-
-    def _sample(self, names, fn, dist, *args, **kwargs):
         if isinstance(names, list):
             names = [self._prefix + n for n in names]
-            lazy_values = [LazyValue(self, k, names, fn, dist, *args, **kwargs) for k in names]
+            lazy_values = [LazyValue(self, k, names, fn, *args, **kwargs) for k in names]
             if self._targets is None or any([k in self._targets for k in names]):
                 lazy_values[0].evaluate()
             return tuple(lazy_values)
         elif isinstance(names, str):
             name = self._prefix + names
-            lazy_value = LazyValue(self, name, name, fn, dist, *args, **kwargs)
+            lazy_value = LazyValue(self, name, name, fn, *args, **kwargs)
             if self._targets is None or name in self._targets:
                 lazy_value.evaluate()
             return lazy_value
@@ -162,19 +154,17 @@ class TracePrefixContextManager:
 class LazyValue:
     """Provides lazy evaluation functionality.
     """
-    def __init__(self, trace, this_name, fn_out_names, fn, dist, *args, **kwargs):
+    def __init__(self, trace, this_name, fn_out_names, fn, *args, **kwargs):
         """Instantiates LazyValue object.
 
         Args:
             trace: Trace instance (to be populated with sample).
             this_name: Name of this the variable that this LazyValue represents.
             fn_out_names: Name or list of names of variables that `fn` returns.
-            fn: Callable or object that (upon instantiation) returns sample or list of samples.
-            dist: Type of fn: Callable (false) or distribution (true)
+            fn: Callable that returns sample or list of samples.
             args, kwargs: Arguments and keyword arguments provided to `fn` upon evaluation.
         """
         self._trace = trace
-        self._dist = dist
         self._this_name = this_name
         self._fn_out_names = fn_out_names
         self._fn = fn
@@ -196,27 +186,15 @@ class LazyValue:
         Returns:
             Value of `this_name`.
         """
-        instance = None
         if self._this_name not in self._trace.keys():
             args = (arg.evaluate() if isinstance(arg, LazyValue) else arg for arg in self._args)
             kwargs = {k: v.evaluate() if isinstance(v, LazyValue) else v for k, v in self._kwargs.items()}
-            if not self._dist:
-                result = self._fn(*args, **kwargs)
[…]
