### 2022-07-14 a218c27
Remove obsolete code
