+++ swyft/lightning/components.py
@@ -1,1125 +1,0 @@
-from abc import abstractmethod
-import math
-from dataclasses import dataclass, field
-from toolz.dicttoolz import valmap
-from typing import (
-    Callable,
-    Dict,
-    Hashable,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-)
-import numpy as np
-import torch
-import torch.nn as nn
-from torch.nn import functional as F
-from torch.utils.data import random_split
-import pytorch_lightning as pl
-from tqdm import tqdm
-import swyft
-import swyft.utils
-from swyft.inference.marginalratioestimator import get_ntrain_nvalid
-import yaml
-
-import zarr
-import fasteners
-from dataclasses import dataclass
-from pytorch_lightning import loggers as pl_loggers
-
-from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint
-from pytorch_lightning import loggers as pl_loggers
-from pytorch_lightning.callbacks.early_stopping import EarlyStopping
-import pytorch_lightning as pl
-from pytorch_lightning.trainer.supporters import CombinedLoader
-
-from swyft.networks.standardization import OnlineStandardizingLayer
-
-
-###########
-# Simulator
-###########
-
-def to_numpy(x, single_precision = False):
-
-    if isinstance(x, torch.Tensor):
-        if not single_precision:
-            return x.detach().cpu().numpy()
-        else:
-            x = x.detach().cpu()
-            if x.dtype == torch.float64:
-                x = x.float().numpy()
-            else:
-                x = x.numpy()
-            return x
-    elif isinstance(x, Samples):
-        return Samples({k: to_numpy(v, single_precision = single_precision) for k, v in x.items()})
-    elif isinstance(x, dict):
-        return {k: to_numpy(v, single_precision = single_precision) for k, v in x.items()}
-    elif isinstance(x, np.ndarray):
-        if not single_precision:
-            return x
-        else:
-            if x.dtype == np.float64:
-                x = np.float32(x)
-            return x
-    else:
-        return x
-
-def to_numpy32(x):
-    return to_numpy(x, single_precision = True)
-    
-def to_torch(x):
-    if isinstance(x, Samples):
-        return Samples({k: to_torch(v) for k, v in x.items()})
-    elif isinstance(x, dict):
-        return {k: to_torch(v) for k, v in x.items()}
-    else:
-        return torch.as_tensor(x)
-    
-
-class Trace(dict):
-    """Defines the computational graph (DAG) and keeps track of simulation results.
-    """
-    def __init__(self, targets = None, conditions = {}):
-        """Instantiate Trace instante.
-
-        Args:
-            targets: Optional list of target sample variables. If provided, execution is stopped after those targets are evaluated. If `None`, all variables in the DAG will be evaluated.
-            conditions: Optional `dict` or Callable. If a `dict`, sample variables will be conditioned to the corresponding values.  If Callable, it will be evaulated and it is expected to return a `dict`.
-        """
-
-        super().__init__(conditions)
-        self._targets = targets
-        self._prefix = ""
-
-    def __repr__(self):
-        return "Trace("+super().__repr__()+")"
-
-    def __setitem__(self, k, v):
-        if k not in self.keys():
-            super().__setitem__(k, v)
-
-    @property
-    def covers_targets(self):
-        return (self._targets is not None 
-                and all([k in self.keys() for k in self._targets]))
-
-    def sample(self, names, fn, *args, **kwargs):
-        """Register sampling function.
-
-        Args:
-            names: Name or list of names of sampling variables.
-            fn: Callable that returns the (list of) sampling variable(s).
[…]
