+++ swyft/lightning/networks.py
@@ -1,0 +1,167 @@
+from abc import abstractmethod
+import math
+from dataclasses import dataclass, field
+from toolz.dicttoolz import valmap
+from typing import (
+    Callable,
+    Dict,
+    Hashable,
+    Optional,
+    Sequence,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
+import numpy as np
+import torch
+import torch.nn as nn
+from torch.nn import functional as F
+from torch.utils.data import random_split
+import pytorch_lightning as pl
+from tqdm import tqdm
+import swyft
+import swyft.utils
+from swyft.inference.marginalratioestimator import get_ntrain_nvalid
+import yaml
+
+import zarr
+import fasteners
+from dataclasses import dataclass
+from pytorch_lightning import loggers as pl_loggers
+
+from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint
+from pytorch_lightning import loggers as pl_loggers
+from pytorch_lightning.callbacks.early_stopping import EarlyStopping
+import pytorch_lightning as pl
+from pytorch_lightning.trainer.supporters import CombinedLoader
+
+from swyft.networks.standardization import OnlineStandardizingLayer
+
+
+##########################
+# Ratio estimator networks
+##########################
+
+def equalize_tensors(a, b):
+    n, m = len(a), len(b)
+    if n == m:
+        return a, b
+    elif n == 1:
+        shape = list(a.shape)
+        shape[0] = m
+        return a.expand(*shape), b
+    elif m == 1:
+        shape = list(b.shape)
+        shape[0] = n
+        return a, b.expand(*shape)
+    elif n < m:
+        assert m%n == 0, "Cannot equalize tensors with non-divisible batch sizes."
+        shape = [1 for _ in range(a.dim())]
+        shape[0] = m//n
+        return a.repeat(*shape), b
+    else:
+        assert n%m == 0, "Cannot equalize tensors with non-divisible batch sizes."
+        shape = [1 for _ in range(b.dim())]
+        shape[0] = n//m
+        return a, b.repeat(*shape)
+    
+# TODO: Introduce RatioEstimatorDense
+class RatioEstimatorMLPnd(torch.nn.Module):
+    def __init__(self, x_dim, marginals, dropout = 0.1, hidden_features = 64, num_blocks = 2):
+        super().__init__()
+        self.marginals = marginals
+        self.ptrans = swyft.networks.ParameterTransform(
+            len(marginals), marginals, online_z_score=False
+        )
+        n_marginals, n_block_parameters = self.ptrans.marginal_block_shape
+        n_observation_features = x_dim
+        self.classifier = swyft.networks.MarginalClassifier(
+            n_marginals,
+            n_observation_features + n_block_parameters,
+            hidden_features=hidden_features,
+            dropout_probability = dropout,
+            num_blocks=num_blocks,
+        )
+        
+    def forward(self, x, z):
+        x, z = equalize_tensors(x, z)
+        z = self.ptrans(z)
+        ratios = self.classifier(x, z)
+        w = Ratios(z, ratios, metadata = {"type": "MarginalMLP", "marginals": self.marginals})
+        return w
+    
+
+class RatioEstimatorMLP1d(torch.nn.Module):
+    def __init__(self, x_dim, z_dim, dropout = 0.1, hidden_features = 64, num_blocks = 2, use_batch_norm = True, ptrans_online_z_score = True):
+        super().__init__()
+        self.marginals = [(i,) for i in range(z_dim)]
+        self.ptrans = swyft.networks.ParameterTransform(
[…]

[…]
