+++ swyft/lightning/__init__.py
@@ -3,3 +3,4 @@ from swyft.lightning.stores import *
 from swyft.lightning.networks import *
 from swyft.lightning.bounds import *
 from swyft.lightning.simulator import *
+from swyft.lightning.samples import *
