+++ swyft/lightning/__init__.py
@@ -1,6 +1,6 @@
 from swyft.lightning.core import *
 from swyft.lightning.stores import *
-from swyft.lightning.networks import *
+from swyft.lightning.estimators import *
 from swyft.lightning.bounds import *
 from swyft.lightning.simulator import *
 from swyft.lightning.samples import *
