+++ swyft/__init__.py
@@ -1,15 +1,21 @@
-from swyft.bounds import Bound
-from swyft.inference import MarginalPosterior, MarginalRatioEstimator
-from swyft.networks import get_marginal_classifier
-from swyft.plot import corner, empirical_z_score_corner, hist1d, violin
-from swyft.prior import (
-    Prior,
-    PriorTruncator,
-    get_diagonal_normal_prior,
-    get_uniform_prior,
-)
-from swyft.store import DaskSimulator, Dataset, Simulator, Store
+#from swyft.bounds import Bound
+#from swyft.inference import MarginalPosterior, MarginalRatioEstimator
+#from swyft.networks import get_marginal_classifier
+#from swyft.plot import corner, empirical_z_score_corner, hist1d, violin
+#from swyft.prior import (
+#    Prior,
+#    PriorTruncator,
+#    get_diagonal_normal_prior,
+#    get_uniform_prior,
+#)
+#from swyft.store import DaskSimulator, Dataset, Simulator, Store
 
+from swyft.lightning.core import *
+from swyft.lightning.stores import *
+from swyft.lightning.estimators import *
+from swyft.lightning.bounds import *
+from swyft.lightning.simulator import *
+from swyft.lightning.samples import *
 
 def zen():
     print("  Cursed by the dimensionality of your nuisance space?")
@@ -19,23 +25,23 @@ def zen():
     print("         Try swyft for some pain relief.")
 
 
-__all__ = [
-    "Bound",
-    "corner",
-    "DaskSimulator",
-    "Dataset",
-    "empirical_z_score_corner",
-    "get_marginal_classifier",
-    "hist1d",
-    "MarginalPosterior",
-    "MarginalRatioEstimator",
-    "OnlineDictStandardizingLayer",
-    "OnlineStandardizingLayer",
-    "Prior",
-    "PriorTruncator",
-    "Simulator",
-    "Store",
-    "get_diagonal_normal_prior",
-    "get_uniform_prior",
-    "violin",
-]
+#__all__ = [
+#    "Bound",
+#    "corner",
+#    "DaskSimulator",
+#    "Dataset",
+#    "empirical_z_score_corner",
+#    "get_marginal_classifier",
+#    "hist1d",
+#    "MarginalPosterior",
+#    "MarginalRatioEstimator",
+#    "OnlineDictStandardizingLayer",
+#    "OnlineStandardizingLayer",
+#    "Prior",
+#    "PriorTruncator",
+#    "Simulator",
+#    "Store",
+#    "get_diagonal_normal_prior",
+#    "get_uniform_prior",
+#    "violin",
+#]
